# web/openapi.json
{
  "components": {
    "schemas": {
...
      "DownloadRequest": {
        "properties": {
          "chapter_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": ["chapter_id"],
        "type": "object"
      },
...
      "JobDto": {
        "properties": {
          "attempts": {
            "format": "int32",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "last_error": {
            "type": ["string", "null"]
          },
          "max_attempts": {
            "format": "int32",
            "type": "integer"
          },
          "priority": {
            "format": "int32",
            "type": "integer"
          },
          "run_at": {
            "format": "date-time",
            "type": "string"
          },
          "state": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "state",
          "priority",
          "run_at",
          "attempts",
          "max_attempts",
          "created_at"
        ],
        "type": "object"
      },
...
      "Paged_JobDto": {
        "description": "One page of results.\n\n`next_cursor` is opaque and is the only supported way to ask for the next\npage. Offsets are deliberately absent: they shift under inserts, and a\nlibrary that gains a chapter mid-scroll would skip or repeat one.",
        "properties": {
          "items": {
            "items": {
              "properties": {
                "attempts": {
                  "format": "int32",
                  "type": "integer"
                },
                "created_at": {
                  "format": "date-time",
                  "type": "string"
                },
                "id": {
                  "format": "uuid",
                  "type": "string"
                },
                "kind": {
                  "type": "string"
                },
                "last_error": {
                  "type": ["string", "null"]
                },
                "max_attempts": {
                  "format": "int32",
                  "type": "integer"
                },
                "priority": {
                  "format": "int32",
                  "type": "integer"
                },
                "run_at": {
                  "format": "date-time",
                  "type": "string"
                },
                "state": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "kind",
                "state",
                "priority",
                "run_at",
                "attempts",
                "max_attempts",
                "created_at"
              ],
              "type": "object"
            },
            "type": "array"
          },
          "next_cursor": {
            "type": ["string", "null"]
          }
        },
        "required": ["items"],
        "type": "object"
      },
...
      "TriggerRequest": {
        "properties": {
          "kind": {
            "description": "One of `update_sources`, `prune_jobs`, `prune_sessions`,\n`reconcile_library`.",
            "type": "string"
          }
        },
        "required": ["kind"],
        "type": "object"
      }
    },
    "securitySchemes": {
      "session": {
        "description": "An opaque, signed, HttpOnly session cookie issued by `GET /api/v1/auth/callback`. There is no bearer token: tokens from the identity provider never reach the browser.",
        "in": "cookie",
        "name": "id",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "A self-hosted manga library server.",
    "license": {
      "name": "AGPL-3.0-or-later"
    },
    "title": "nyuka",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/downloads": {
      "post": {
        "description": "Returns `202` with a `Location` pointing at the job. The work is a series\nof requests to a third-party site; doing it inline would hold a connection\nopen for minutes.\n\nIdempotent on the chapter, so a double-click queues one download. An\n`Idempotency-Key` header is accepted and narrows that further, which is\nwhat makes a retried request from a flaky network safe.",
        "operationId": "requestDownload",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Queued; `Location` names the job"
          },
          "404": {
            "description": "No such chapter"
          },
          "409": {
            "description": "Already in the library"
          }
        },
        "summary": "`POST /api/v1/downloads`",
        "tags": ["downloads"]
      }
    },
...
    "/follows/{id}/check-now": {
      "post": {
        "operationId": "checkFollowNow",
        "parameters": [
          {
            "description": "Follow id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "A check was queued"
          },
          "404": {
            "description": "No such follow"
          }
        },
        "summary": "`POST /api/v1/follows/{id}/check-now`",
        "tags": ["follows"]
      }
    },
    "/jobs": {
      "get": {
        "operationId": "listJobs",
        "parameters": [
          {
            "description": "One of `queued`, `running`, `succeeded`, `failed`, `cancelled`.",
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "type": ["string", "null"]
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": ["string", "null"]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paged_JobDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "description": "Unknown state filter"
          }
        },
        "summary": "`GET /api/v1/jobs`",
        "tags": ["jobs"]
      },
      "post": {
        "description": "Exists because the alternative is a runbook that tells an operator to\n`INSERT` into the `job` table. A documented raw insert is a schema\ndependency in prose: it survives exactly until a column changes, and it\nbypasses every validation the enqueue path performs.",
        "operationId": "triggerJob",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriggerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Queued; `Location` names the job"
          },
          "400": {
            "description": "Not a kind that can be triggered by hand"
          }
        },
        "summary": "`POST /api/v1/jobs` — queue a maintenance job.",
        "tags": ["jobs"]
      }
    },
    "/jobs/{id}": {
      "get": {
        "operationId": "getJob",
        "parameters": [
          {
            "description": "Job id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": "No such job"
          }
        },
        "summary": "`GET /api/v1/jobs/{id}`",
        "tags": ["jobs"]
      }
    },
    "/jobs/{id}/cancel": {
      "post": {
        "description": "Cancelling a `running` job stops it from being retried; it does not\ninterrupt the attempt in flight. That is an honest limit rather than an\noversight — see ADR-0003's open question on cancellation semantics.",
        "operationId": "cancelJob",
        "parameters": [
          {
            "description": "Job id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": "No such job"
          },
          "409": {
            "description": "The job has already finished"
          }
        },
        "summary": "`POST /api/v1/jobs/{id}/cancel`",
        "tags": ["jobs"]
      }
    },
    "/jobs/{id}/retry": {
      "post": {
        "operationId": "retryJob",
        "parameters": [
          {
            "description": "Job id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Queued again"
          },
          "404": {
            "description": "No such job"
          },
          "409": {
            "description": "The job has not finished"
          }
        },
        "summary": "`POST /api/v1/jobs/{id}/retry`",
        "tags": ["jobs"]
      }
    },
...
  },
...
}